requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39h3811e60_0
aiosignal=1.2.0=pyhd8ed1ab_0
ale-py=0.7.4=pypi_0
alsa-lib=1.2.3=h516909a_0
antlr4-python3-runtime=4.8=pypi_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
autorom=0.4.2=pypi_0
autorom-accept-rom-license=0.4.2=pypi_0
blas=1.0=mkl
blinker=1.4=py_1
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.10.8=py39hf3d152e_2
cffi=1.15.0=py39hd667e15_1
cfgv=3.3.1=pyhd8ed1ab_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyh9f0ad1d_0
cloudpickle=2.0.0=pyhd8ed1ab_0
configparser=5.2.0=pyhd8ed1ab_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=ha36c431_10
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=h48d8840_2
distlib=0.3.4=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
expat=2.4.8=h27087fc_0
ffmpeg=4.3.2=h37c90e5_3
filelock=3.6.0=pyhd8ed1ab_0
fontconfig=2.13.96=h8e229c2_2
fonttools=4.31.2=py39hb9d737c_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=py39h3811e60_0
gettext=0.19.8.1=h0b5b191_1005
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.44.0=pypi_0
gst-plugins-base=1.18.5=hf529b03_0
gstreamer=1.18.5=h76c114f_0
gym=0.23.0=pypi_0
gym-notices=0.0.6=pypi_0
harfbuzz=2.9.1=h83ec7ef_1
hdf5=1.10.6=nompi_h7c3c948_1111
icu=68.2=h9c3ff4c_0
identify=2.4.12=pyhd8ed1ab_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39hf3d152e_0
importlib-resources=5.6.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jasper=1.900.1=h07fcdf6_1006
jpeg=9d=h7f8727e_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.2=py39hf939315_0
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libcurl=7.82.0=h7bff187_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h9b69904_4
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libglib=2.68.4=h3e27bee_0
libgomp=11.2.0=h1d223b6_14
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
liblapacke=3.9.0=12_linux64_mkl
libllvm11=11.1.0=hf817b99_3
libnghttp2=1.47.0=h727a467_0
libogg=1.3.4=h7f98852_1
libopencv=4.5.1=py39hf3c14e0_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libpq=13.5=hd57d9b9_1
libprotobuf=3.20.0=h6239696_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h166bdaf_1014
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib-base=3.5.1=py39h2fa2bec_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=py39h3811e60_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.25=ha770c72_0
mysql-libs=8.0.25=h935591d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
nodeenv=1.6.0=pyhd8ed1ab_0
nspr=4.32=h9c3ff4c_1
nss=3.76=h2350873_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.2.0=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.1.1=pypi_0
opencv=4.5.1=py39hf3d152e_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=pypi_0
pathtools=0.1.2=py_1
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h36c2ea0_0
platformdirs=2.5.1=pyhd8ed1ab_0
pre-commit=2.17.0=py39hf3d152e_0
promise=2.3=py39hf3d152e_5
protobuf=3.20.0=py39h5a03fae_0
psutil=5.9.0=py39h3811e60_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.1=py39hef51801_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygame=2.1.2=pypi_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39h3811e60_3
qt=5.12.9=hda022c4_4
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scipy=1.7.3=py39hc147768_0
sentry-sdk=1.5.8=pyhd8ed1ab_0
setproctitle=1.2.2=py39h3811e60_1
setuptools=58.0.4=py39h06a4308_0
shortuuid=1.0.8=py39hf3d152e_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
sqlite=3.38.0=hc218d9a_0
stable-baselines3=1.1.0=pyhd8ed1ab_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
termcolor=1.1.0=py_2
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchvision=0.12.0=py39_cu113
tqdm=4.62.3=pyhd3eb1b0_1
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2021e=hda174b7_0
ukkonen=1.0.1=py39h1a9c180_1
unicodedata2=14.0.0=py39h3811e60_0
urllib3=1.26.8=pyhd3eb1b0_0
virtualenv=20.14.0=py39hf3d152e_0
wandb=0.12.11=pyhd8ed1ab_0
werkzeug=2.1.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39h3811e60_1
yaspin=2.1.0=pyhd8ed1ab_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h166bdaf_1014
zstd=1.4.9=haebb681_0
